+++ src/ItemsDialog.test.js
@@ -8,6 +8,7 @@ import {
   MultiColumnList,
   Pane,
   Paneset,
+  NoValue,
 } from '@folio/stripes/components';
 
 import {
@@ -392,6 +393,7 @@ describe('ItemsDialog', () => {
 
   describe('formatter', () => {
     const item = {
+      barcode: 'barcode',
       status: {
         name: 'Aged to lost',
       },
@@ -409,6 +411,25 @@ describe('ItemsDialog', () => {
       },
     };
 
+    afterEach(() => {
+      NoValue.mockClear();
+    });
+
+    describe('barcode', () => {
+      it('should return item barcode', () => {
+        expect(formatter.barcode(item)).toBe(item.barcode);
+      });
+
+      it('should trigger NoValue component', () => {
+        render(formatter.barcode({
+          ...item,
+          barcode: undefined,
+        }));
+
+        expect(NoValue).toHaveBeenCalled();
+      });
+    });
+
     describe('itemStatus', () => {
       it('should return formatted message', () => {
         expect(formatter.itemStatus(item).props.id).toEqual('ui-requests.item.status.agedToLost');
@@ -420,11 +441,13 @@ describe('ItemsDialog', () => {
         expect(formatter.location(item)).toEqual('effective location name');
       });
 
-      it('should return default value for effective location name', () => {
-        expect(formatter.location({
+      it('should trigger NoValue component', () => {
+        render(formatter.location({
           ...item,
           effectiveLocation: {},
-        })).toEqual('');
+        }));
+
+        expect(NoValue).toHaveBeenCalled();
       });
     });
 
@@ -440,13 +463,15 @@ describe('ItemsDialog', () => {
           expect(formatter.loanType(item)).toEqual('temporary loan type name');
         });
 
-        it('should return default value for temporary loan type name', () => {
-          expect(formatter.loanType({
+        it('should trigger NoValue component', () => {
+          render(formatter.loanType({
             ...item,
             temporaryLoanType: {
               other: '',
             },
-          })).toEqual('');
+          }));
+
+          expect(NoValue).toHaveBeenCalled();
         });
       });
 
@@ -458,14 +483,16 @@ describe('ItemsDialog', () => {
           })).toEqual('permanent loan type name');
         });
 
-        it('should return default value for permanent loan type name', () => {
-          expect(formatter.loanType({
+        it('should trigger NoValue component', () => {
+          render(formatter.loanType({
             ...item,
             temporaryLoanType: false,
             permanentLoanType: {
               other: '',
             },
-          })).toEqual('');
+          }));
+
+          expect(NoValue).toHaveBeenCalled();
         });
       });
     });